Limits scope of "space hiding skeleton" key press.
This removes the global event listener from the page and places a
keypress event listener on the results section. Now, the results have to
be in focus, "clicking any where in the div", for the spacebar to
trigger the event that closes the skeleton view. This also stops the
nasty bug of not being able to type a ' ' in the query parameters.