Fixes neuroglnacer css and responsive grid scrolling
This missing <!DOCTYPE html> tag at the top of the html document was
causing the browser to render in quirks mode. This broke the flexbox
layout used by neuroglancer. Adding the doctype tag broke the window
scrolling for the responsive grid layout component, because there was
an overflow:hidden style applied to the body tag. Removing that seems to
have put things back on track.