Adds a new plugin system.
The new system reduces the number of files needed to create a query
plugin, by splitting the fetching and rendering of the results into
separate generic plugins. The plugins directory contains plugin files
that are responsible for fetching, formatting and storing a query and
the results. The view-plugins directory contains plugins that expect
data in a specific format and know how to take that data and render it
into a react component.