Adds Skeleton actions and reducers.
This hooks the Skeleton viewer into the Testplugin via actions that
send data to the redux store. The store is then checked by the Results
component to see if the skeleton needs to be rendered. If it does, the
Skeleton component is added to the page and loops over the stored data
to generate the 3D rendering.
This change makes it easier to call the Skeleton viewer from any plugin.