Updates the way url query parameters are set.
It now uses the central history object used by react router to make sure
that we are creating a consistent history. Secondly the query parameter
no longer has to be pulled into the object, manipulated and then saved
back out. We can simply save the data structure that we want and it will
get merged with the existing one and pushed out to the url.