Fixes neuroglancer updates and shark viewer.
The neuroglancer window was not updating when new neurons were being
selected. This was because the react component didn't have an update
mechanism. Now that we are including the updated version of the react
component, it does. Shark viewer was always reinitialising the canvas on
every update, even if the number of neurons didn't change, now it
doesn't.