Converts spacebar toggle of neuron viz
The keyPress event was attached to the div containing the results in an
attempt to limit the scope of the event listener. This ended up causing
an a11y issue, due to the div not being an interactive element. The fix
is to move the handler to the document and only trigger the effect if
the mouse is hovering over the parent div. The hover detection is done
using the onMouseEnter and onMouseLeave events.