Removes hovered state handlers
These were causing the Results component to re-render every time the
mouse entered or left the results component. This was horrible for
performance, so the handlers have been disabled. The next step is
to figure out how to change the state without forcing a re-render. This
most likely route is to extract the hover check into its own component.