Cleans up tab behavior when adding/removing neuronviz
This fixes blank page issues when removing the skeleton or neuroglancer
tabs. Now when the tabs are removed the first tab in the list is
selected and displayed. It also jumps to the neuroglancer tab when an
icon to show the skeleton is clicked, rather than just creating the tab
and not showing it.