Now storing camera position in application state.
This happens when the skeleton component is unmounted. Then if it is
remounted, the state is checked and the position is applied to the new
camera object. This makes it possible to exit the tab and return to the
same view state.

The reason the position is only saved when the component unmounts is
that there was a performance problem when trying to update the state
every time the camera moved. This caused lagging redraws and redux to
fail. So storing it on unmount is a compromise to get the functionality
we need.