Fixes blank canvas issue on component mount.
There were a number of cases where the canvas would be blank when the
Skeleton component was remounted. It seems that the componentDidMount
function was being too aggressive when deciding not to reload all the
neurons back into the scene. This update removes those checks and seems
to remove the blank canvas bug.