Fixes camera position issues in skeleton viewer.
There were a number of cases where the skeleton view was not being
restored to the previous view. These cases were being caused by a change
in the camera target, which we were not tracking. This update adds the
target to the values we are tracking and fixes the bugs.