Adds pouchDB for compartment data caching.
The ROI obj files are too large to store in the state as they cause
rendering updates to crawl to a halt. They are also too large to store
in localStorage, which is limited to ~5Mb. The solution is to use
indexDB or a wrapper around it. I chose pouchDB as it is crossbrowser
and works much like couchDB.