Only pass ROIs and neurons to load if changed.
There was a timing issue that was causing ROIs to load twice when the
skeleton viewer component was mounted into the page. This was caused by
async loading of the components and the wrapping code not knowing when a
load was in progress. Now the update code checks to see if the neurons
or ROIs have changed since the component mounted and only sends the ones
that have to be rendered.