Fixes pagination and size settings in SimpleTable.
The handlers that used to update page size or current page for a results
tab were using the redux state to keep track of their value. This
changes the code to use the url query string to keep track of the
current page and size. That way it survives a cut & paste of the url or
a refresh of the browser.