Adds index to ensure skeletons removed from correct tab
The skeletons were not always being removed from the correct tab with
the previous code, because it had to guess which tab was the intended
target. This was fine when there was only one tab for each dataset, but
when there are multiple tabs, it can get it wrong. Passing the tab index
into the remove function removes the guesswork.