Adds heatmap button css.
This adds the css definition to the results component, because if we
define it in the plugin, the css isn't applied to the simple table
result when the query plugin is unmounted. This is not ideal, because
the css definition is not with the plugin, but I can't think of
another way around it currently.