Adds Error Catching to Results component.
This adds some code to trap errors from the plugins. Ideally it should
prevent the application from crashing if a plugin fails to parse the
results correctly or the server responds with an error. The messages
that are displayed are a bit too cryptic at the moment, but that can be
improved with some better error handling code.