Fixed caching of results for tabs with same plugin
The caching code was only checking to see if the plugin code in the
cached results matched the code from the query parameters. This was not
sufficient to prevent errors, such as when two tabs use the same plugin,
but different datasets. This update splits the query parameters and the
result object and stores them both in the redux state. Then the entire
parameter object is compared to the cached one to see if we need to load
new results.