Clone cached result before passing to plugin.
If we don't do this, then the plugin can permanently modify the cached
data. This would be fine if it was loading once, but when it gets loaded
more than once, the effects of the processing can be multiplied. For
example:

const results = [0,1,2,3];
// processResults adds an extra value to the end of the results array.
const processed = plugin.processResults(results);
// processed === [0,1,2,3,none];
// results === [0,1,2,3,none];
const processed_again = plugin.processResults(results);
// processed_again === [0,1,2,3,none,none];
// results === [0,1,2,3,none,none]; <- bad!!