PLUGIN_SUBMITTING now clears result data.
Previously it would remove the entry from the allResults array, which
would cause all the cached results for the other tabs to be out of
place. By nulling the data in the indexed position, we keep the order
correct, but also make sure that a cache hit is not possible for the tab
while loading new data.