Increases query string arrayLimit.
The qs library set a limit of 20 for the number of items in an array
that is passed to it. This limit would convert any array over a length
of 20 to an object. This resulted in malformed queries to the API, which
was expecting an array, causing the request to fail. The limit has been
bumped up to 2000 and we will need to add some validation to the form to
make sure more than that are not submitted.