Adds Saved search display.
Changes have been made to support the new saved search plugin, which
will take the saved data from a saved search and display it as it would
have been displayed before it was saved. This enables people to save a
set of search results and compare them to the same search at a later
date.