Corrects visProps passes to results processing.
The incorrect visProps were being passed to the processResults() when
restored from a saved search. This caused the resulting display to look
incorrect. Now the data in the page looks correct, but sometimes the row
numbers per page are off, because the visprops need to be modified later
in the Results component.