Adds cached time string to results tab header.
This shows the length of time since the contents of the current results
tab have been loaded from the server. Refreshing the tab or the whole
page will reset this value. It is also in a constantly updating human
readable format to make it easier to figure out how stale the data is.

'Loaded from server 2 minutes ago'

is easier to parse than

'Loaded: 10/07/2019 10:34'