Fixes infinite loading loop on result load error.
The symptoms of this bug were first identified when, custom query plugin
failed to load, because of a bad cypher query. The error message would
show and it was impossible to dismiss it. This was because the state
was being checked to see if there was a value in the results data
structure. If there was no value, as in the case of a load error, the
page was being reloaded to try and get a value. This bug was introduced
when the refresh button was added. So, now the code checks to see if the
results data is empty and if there is an error message present. If there
is an error, then the reload is not fired, to prevent the infinite loop.
Unfortunately, to make this work correctly we have to convert the
loadingError state from a string to an Immutable.List. This allows us to
store error messages on a per tab basis, so that an error on one tab
does not affect the loading of another tab.