Fixes loading of saved result for skeleton views.
The skeleton viewer plugin was not loading the saved results, because
the information it needed was not being passed into the view component.
The information was being replaced by the information for the saved
search component, which meant the skeleton viewer had no way to find the
body ids it needed to load.