Moves SkeletonView plugin back into the core code.
This makes it much easier to access the redux state, which is necessary
to store the selected synapse information, since it is too extensive to
store in the url query string. It will also make it easier to save the
state to a data store and restore it when people wish to share a link to
the results they are looking at.