Changes alignment of body id chips in SkeletonView
The alignment has moved from a horizontal row to a vertical column down
the left side of the view. This stops the ROI input box rendering over
the top of the chips, if there are a lot of them. It does introduce a
problem with the chips not being accessible if the screen is really
short, but that seems to be less of an issue than the ROI inputs. Future
work is probably needed here.