Clears out results cache when query parameters are cleared
Clear out the cache if we no longer have the query result parameters in the url.
This prevents strange errors, eg: clicking on the heatmap on the homepage with a
cached result could cause the site to crash as it would try and use a cached
result for the wrong ROI. Or if there were 10 results in the cache and the url
was cleared of the result parameters, there is no way to start a new search,
because the site says there are too many results, but there is no way to clear
them without refreshing the page. Currently only clears out the results
cache when loading the home page component as that is the only place
that has been identified as needing the cache clearing.