First pass at getting unique tab labels.
Adds a number to each tab to try and make it easier to identify if
multiple tabs of the same type are open. Does not work as well as hoped,
as it it possible to have multiple tabs with the same number and name
displayed. This happens when for example, a tab is opened at position 3
and then the tab at position 1 is closed and another tab is opened. It
will be assigned the number 3 and the same plugin name, resulting in the
same label. Need to put more thought into this.