Separates Results tabs from result content.
There was a single Results component that handled loading in the data
for the currently selected tab, but it also rendered the tabs. This
keeps the tab rendering in the Results component and moves the content
rendering into the Result component. This is a step towards having
multiple Result components rendering at the same time. To demonstrate
that, the current code has tab '0' showing at all times, in addition to
the other tab that is selected. This will need to be changed so that we
can choose which tabs to display in the split panel.