Fixes render bug in ROICompletenessChart
In some browsers the loading of the superROI data happens after the
loading of the ROICompleteness info. This causes a bug, because the data
is undefined. The fix is to store the data and only apply the filter
when the superROIs have loaded. This is annoying as the filter has to be
applied every time the component is loaded, but it is more reliable and
doesn't fail due to timing issues.