Adds papaparse csv parser for handling downloads.
There are a lot of edge cases when dealing with CSV files that we
weren't handling. This library deals with those and more, allowing use
to create an array of arrays and pass it in to get a properly formatted
CSV file. See https://www.papaparse.com/ for more details.