Skeleton viewer camera movement prevented.
The skeleton viewer camera would jump to look at the center point of any
new bodies added to the page, unless there were coordinates found in the
url. This update also prevents the camera move if there is already a
body in the scene. The result is the camera will only move to the center
point of the first loaded body and resets on that body whenever the
reset button is pressed.