Updates webpack config to find node_module changes
The changes being made in the @neuprint modules were not being picked up
and applied by the webpack server after the migration to webpack 5. This
meant reloading the whole development server after each change to check.
With this change the development server now picks up the changes and
the website can be updated without reloading the development server each
time.