Fixes addition of single bodies to skeleton view
When the code to add multiple bodies was added to the skeleton reducer,
it inadvertently caused the addition of single bodies to fail. This is
because the single bodies were no longer in the correct format. It also
caused the addition to fail, because it created empty color lookup
records in the pouchdb cache.