feat: Find Neurons columns configured per dataset
The neuprintHTTP server can now provide the n.neuronColumns and
n.neuronColumnsVisible meta attributes. These are loaded by the MetaInfo
component of neuprintExplorer and can be accessed through the
neo4jsettings.ColumnDefaults key in the redux store. Using these values
we can merge the columns configured in the FindNeurons component with
the defaults sent from the server to create a new list of columns
customised to each dataset.