fix: Adds maxVolumeSize to configure SkeletonView
Each dataset will have an varying maximum volume size for the neurons
that are displayed in the SkeletonView. Adding this parameter to the
datasetInfo values returned by neuprint allows the dataset owner to
change the maxVolumeSize for a specific dataset and allow the user to
zoom out more or less, based on that volume size. The default volume
size is 100000 pixels.