feat: Returns plugins to this repo.
This deprecates the neuprintExplorerPlugins repo and moves all the code
that was hosted there back into this repo as components that get loaded
directly by the core code. This makes development much easier and
reduces the cognitive load when making changes to the site. The initial
idea behind the plugins was to allow contributors to contribute to a
separate repo and have their own plugins for the server. This feature
has never been used as far as I know. Returning this code to this repo
makes sense for the small group of developers making changes here.