Feat: Adds account switch for pinned viewer choice
The logged in user can choose to have the skeleton view as the default
viewer, or they can have neuroglancer. This is so that people who don't
like the move to using neuroglancer as the default viewer, can go back
to having the skeleton view that they are used to.