feat: Adds 'Pin' icon to results top bar.
The pin to side button allows someone to pin any of the panels off to
the side, in the same way that the neuroglancer or skeleton views have
been when opened from a bodyId link. Now the power is in the users hands
to decide what should be in the side panel.