fix: Fixes addition of bodies to 3d viewers.
There were a number of bugs when adding a new body to one or both of the
3d viewers, which lead to some of the bodies not being displayed in
neuroglancer or some of them not being displayed in the skeleton viewer.
This is a large rewrite of the reducer that handles these events. It
also resulted in converting the neuroglancerview component into a
functional component.