fix: Replaces connectivity heatmap with @nivo/heatmap
The original custom d3 svg heatmap was too slow to load when the number
of cells got too large. The CNS dataset had 12,000 cells and stopped the
page from being interactive for over 5 seconds. The heatmap component
from https://nivo.rocks/ is faster to load and shows the same data in a
responsive display.