fix: Neuroglancer actions moved to {dataset}-actions.json
The neuroglancer actions configuration for the neuprint menus has been
moved out of the nglayers file and into a standalone file that gets
loaded by the menu bar. This improves the developers experience,
allowing them to simply dump a neuroglancer state into the original
nglayers file and not have to worry about merging it with actions.
Thanks for the reasonable improvement suggestion, @stuarteberg