fix: Improves performance of the neuroglancer view
    Switching the default 3D viewer for the site to neuroglancer introduced
a new issue with having neuroglancer pinned to the side of a results
section. The problem was that the neuroglancer view component would get
rerednered on every change to the url, which included adding additional
neurons to the view from the results tables. This caused a very slow
loading of all the data in the neuroglancer window and was making the
view unusable for some.

    To fix this, the context that stores the state of the neuroglancer
view was moved above the results component, so that it could maintain
state, even when the results component was updated. In addition to this,
the functions that add a body to the 3D view were made aware of the
viewer that is visible at the time of the click event. If it is the
neuroglancer viewer, then the data is added to the neuroglancer context
and the url is updated using a window.history.replaceState() call. If
the view is the skeleton viewer, then the url is updated using the more
typical history.push() call.