fix: Prevents extraneous type & instance options.
This prevents the case where the input text matches the synonyms field,
but doesn't match anywhere, and it is shown as an instance and type in
the options. eg: If someone types in 10013, they want to see the MBON01
types and instances, but not the instances and types that are related to
the synonym that also has 10013 in the text. That synonym is shown in
the synonyms section, but doesn't need to have it's type or instance
shown in those sections.