refactor: Moves loading of logo & description to MetaInfo component.
The MetaInfo component is called when the site loads and is responsible
for loading all of the meta information about a dataset. This seemed
like the perfect place to put the loading of the logo and description
for each data set. Doing that meant that loading that information could
be removed from the DataSetLogo component, reducing the number of
queries required to get the page loaded.